requirements.txt
psycopg2-binary==2.9.9
mysql-connector-python==8.3.0
PyYAML==6.0.1
sqlparse==0.4.4
rich==13.7.0
